new & fix : UserDetails 상속을 표현하기 위한 클래스 이름 변경 및 재설계
