--- src/main/java/nextstep/member/application/LoginMemberService.java
@@ -1,20 +1,0 @@
-package nextstep.member.application;
-
-import nextstep.member.domain.LoginMember;
-import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberRepository;
-import org.springframework.stereotype.Service;
-
-@Service
-public class LoginMemberService {
-    private MemberRepository memberRepository;
-
-    public LoginMemberService(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    public LoginMember loadUserByUsername(String email) {
-        Member member = memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
-        return LoginMember.of(member);
-    }
-}
